+++ src/caad_erp/cli/commands/restock.py
@@ -1,5 +1,4 @@
 import argparse
-from decimal import Decimal
 
 from caad_erp import bll
 
@@ -57,8 +56,8 @@ def _translate_restock(args: argparse.Namespace) -> bll.RestockCommand:
     return bll.RestockCommand(
         product_id=args.product_id,
         salesman_id=args.salesman_id,
-        quantity=Decimal(args.quantity),
-        total_cost=Decimal(args.total_cost),
+        quantity=int(args.quantity),
+        total_cost=int(args.total_cost),
         notes=args.notes,
     )
